+++ packages/api/src/constants.ts
@@ -1,4 +1,0 @@
-export const DEFAULT_SORT = <const>{
-  created_at: -1
-}
-