+++ packages/api/src/index.ts
@@ -1,7 +1,6 @@
 export * from './assets'
 export * from './collection'
 export * from './context'
-export * from './constants'
 export * from './token'
 export * from './types'
 export * from './database'